# openapi.yaml
openapi: 3.0.3
info:
  title: Panda Market API
  version: 1.0.0
servers:
  - url: http://localhost:3001
...
paths:
  /auth/signUp:
    post:
      tags: [Auth]
      description: 이메일 회원가입 및 JWT 발급
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/signIn:
    post:
      tags: [Auth]
      description: 이메일 로그인 및 JWT 발급
      operationId: signIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/refresh-token:
    post:
      tags: [Auth]
      description: 리프레시 토큰으로 액세스 토큰 재발급
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: 재발급 성공
...
  /auth/google/callback:
    get:
      tags: [Auth]
      description: 구글 OAuth 콜백을 처리하고 클라이언트로 이동합니다.
      operationId: finishGoogleOAuth
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: 인증 토큰과 함께 클라이언트로 이동
...
  /articles:
    post:
      tags:
        - Article
      description: 게시글 생성 API
      operationId: createArticle
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleRequest'
      responses:
        '201':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          description: Validation failed

    get:
      tags:
        - Article
      description: 게시글 목록 조회 API
      operationId: getArticleList
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
          example: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 20
          example: 10
        - name: orderBy
          in: query
          required: false
          schema:
            type: string
            enum:
              - recent
              - favorite
          example: recent
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          example: "검색어"
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetArticleListResponse'
  /articles/{articleId}:
    get:
      tags:
        - Article
      description: 게시글 조회 API
      operationId: getArticle
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '404':
          description: Not found

    patch:
      tags:
        - Article
      description: "게시글 수정 API"
      operationId: updateArticle
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '400':
          description: Validation failed
        '404':
          description: Not found

    delete:
      tags:
        - Article
      description: "게시글 삭제 API"
      operationId: deleteArticle
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '204':
          description: 성공
        '404':
          description: Not found
  /articles/{articleId}/like:
    post:
      tags: [Article]
      description: 게시글 좋아요
      operationId: likeArticle
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArticleId'
      responses:
        '201':
          description: 좋아요 성공
    delete:
      tags: [Article]
      description: 게시글 좋아요 취소
      operationId: unlikeArticle
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArticleId'
      responses:
        '201':
          description: 좋아요 취소 성공
  /articles/{articleId}/comments:
    post:
      tags:
        - Article
      description: 댓글 생성 API
      operationId: createArticleComment
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleCommentRequest'

      responses:
        '201':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Validation failed
        '404':
          description: Not found

    get:
      tags:
        - Article
      description: 댓글 목록 조회 API
      operationId: getArticleCommentList
      parameters:
        - name: articleId
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: cursor
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
          example: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 10
          example: 10
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetArticleCommentListResponse'
        '404':
          description: Not found'
...
  /comments/{commentId}:
    patch:
      tags:
        - Comment
      description: "댓글 수정 API"
      operationId: updateComment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Validation failed
        '404':
          description: Not found

    delete:
      tags:
        - Comment
      description: "댓글 삭제 API"
      operationId: deleteComment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '204':
          description: 성공
        '404':
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ArticleId:
      name: articleId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
    SignUpRequest:
      type: object
      required: [email, nickname, password, passwordConfirmation]
      properties:
        email:
          type: string
          format: email
        nickname:
          type: string
        password:
          type: string
          format: password
        passwordConfirmation:
          type: string
          format: password
    SignInRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        nickname:
          type: string
        image:
          type: string
          nullable: true
    CreateArticleRequest:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          example: "게시글 제목"
        content:
          type: string
          example: "게시글 내용"
        image:
          type: string
          format: url
          nullable: true
          example: "http://localhost:13000/static/images/sample-image.png"
    GetArticleListResponse:
      type: object
      properties:
        totalCount:
          type: integer
          example: 42
        list:
          type: array
          items:
            $ref: '#/components/schemas/Article'
    UpdateArticleRequest:
      type: object
      properties:
        title:
          type: string
          example: "게시글 제목"
        content:
          type: string
          example: "게시글 내용"
        image:
          type: string
          format: url
          nullable: true
          example: "http://localhost:13000/static/images/sample-image.png"
    CreateArticleCommentRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          example: "댓글 내용"
    GetArticleCommentListResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        nextCursor:
          type: integer
          nullable: true
          example: 10
        hasNext:
          type: boolean
          example: true
...
    UpdateCommentRequest:
      type: object
      properties:
        content:
          type: string
          example: "댓글 내용"
    Article:
      type: object
      properties:
        id:
          type: integer
          example: 1
        writer:
          type: object
          properties:
            id:
              type: integer
            nickname:
              type: string
        title:
          type: string
          example: "게시글 제목"
        content:
          type: string
          example: "게시글 내용"
        image:
          type: string
          format: url
          nullable: true
          example: "http://localhost:13000/static/images/sample-image.png"
        createdAt:
          type: string
          format: datetime
          example: "2024-07-15T00:00:00.000Z"
        favoriteCount:
          type: integer
          example: 3
        isLiked:
          type: boolean
          example: false
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: datetime
          example: "2024-07-15T00:00:00.000Z"